Add files via upload
Therapy chatbots are a promising solution to increase access to mental health support. Machine learning (ML) techniques can be used to develop therapy chatbots that can simulate human conversations and provide personalized support to individuals. In this context, ML models are trained on large amounts of text data from therapy sessions to learn how to generate appropriate responses to a user's inputs. The architecture of a therapy chatbot typically consists of several components, including a natural language processing (NLP) 
module. The NLP module processes the user's input and extracts relevant information, such 
as the user's emotional state or the topic they want to discuss. Then the module determines 
the appropriate response based on the user's input and the chatbot's knowledge base. Finally, 
it can suggest additional resources or interventions, such as exercises or referrals to human 
therapists, based on the user's needs. It provides the necessary output based on the type of 
input. For instance, when we provide text as input, we receive text as output, and when we 
provide voice as input, we receive both text and voice as output.
Preliminary evaluations indicate that the therapy chatbot can effectively reduce symptoms 
of anxiety and depression, and improve overall emotional well-being. While it is not a 
substitute for professional therapy, it serves as a valuable supplementary tool, particularly 
for those who face barriers to accessing traditional mental health services